requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
sqlalchemy==2.0.49
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic[email]==2.13.3
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
httpx==0.28.1
anthropic==0.97.0
